Fix the log levels mapping
The libssh provides the most verbose logging with SSH_LOG_TRACE, which
was not mapped to any of the standard values so the users are unable
to get full debug logs. These are critical for libssh developers to be
able to investigate issues.

This also creates callback for libssh to feed the logs into the python
logging system to be able to process the logs the python way.

The conftest no longer needs setting log level as it is set to the
highest level by default and filtered only on the python side.

Signed-off-by: Jakub Jelen <jjelen@redhat.com>
Co-authored-by: Sviatoslav Sydorenko <webknjaz@redhat.com>