channels: Avoid reusing stack-local variables for callbacks
Previously, stack-local variables held the callback structure.
This worked as long as the call stack was not overridden by other
function calls or as long as the server closed the channel
very early. In other cases, the delayed close in libssh
could result in calling the callbacks even after we believed
the channel was freed, causing invalid memory access and crashes.

The accompanying change to libssh was merged to avoid calling
callbacks on channels the caller considers freed here:

https://gitlab.com/libssh/libssh-mirror/-/merge_requests/549/

But we will be using older libssh versions for some time so we
need a workaround in pylibssh too.

Fixes #57

Signed-off-by: Jakub Jelen <jjelen@redhat.com>