tests: Log all sshd messages to stderr
When the OpenSSH server runs in daemon mode, the logs are written into
syslog. This is not suitable for tests as we can not see what is going
on there. The captured logs are only from the main sshd process and
does not show any useful information about the connection and what we
actually test.

The trick to get debug logs is either to not run in deamon mode, but it
would accept only one connection, or use the `-E` to write logs from all
connections to the given file.

Signed-off-by: Jakub Jelen <jjelen@redhat.com>