channel: Wait for async messages before closing channel
Surprisingly, the SSH protocol does not define if the exit-code
notification can come before or after the data. Depending on the server,
scheduler or current mood, the exit-code might arrive earlier than the
data and if we close the channel right away, we will lose some output.

This changes the code to correctly wait for the EOF before closing the
channel and returning.

Signed-off-by: Jakub Jelen <jjelen@redhat.com>