🧪 Bump actions/{upload,download}-artifact to v4
They are updated across all GitHub Actions CI/CD workflow definitions
in the repository. It is necessary due to Artifacts API v3 having been
deprecated on Jan 30, 2025.

As a part of the change, the uploaded artifact names needed to be made
unique and their download counterparts had to be invoked against a
name pattern, merging multiple artifacts together.

Making the names unique came with challenges that this patch overcame
by appending SHA-256 hashes of reusable workflow inputs to the
artifact names.

[1]: https://github.blog/changelog/2024-04-16-deprecation-notice-v3-of-the-artifact-actions/

Co-Authored-By: Sviatoslav Sydorenko <webknjaz@redhat.com>

Signed-Off-By: NilashishC <nilashishchakraborty8@gmail.com>
Signed-Off-By: Sviatoslav Sydorenko <webknjaz@redhat.com>