🐛📦 Fix the OCI manylinux image platform tag
This was caused by a mistake during the initial migration from docker
to podman in #181. The original command looked like
`docker buildx build --platform linux/arm64` but the
`redhat-actions/buildah-build` action has an input called `arch:` that
we started using. And we've prefixed the passed value with `linux/`.
It would've been fine if we used the `platform:` input but we didn't.
The consequence was that the images we were making were tagged with an
additional leading `linux/` prefix in the platform metadata which made
it look like `linux/linux/arm64`.

The issue became evident in #648 that attempted to bump the version of
`cibuildwheel`. And the container interaction started failing loudly
as this tool started using the `--platform` option when working with
OCI images in v2.21 [[1]].

[1]: https://github.com/pypa/cibuildwheel/pull/1961