Update OpenSSL to 3.5.0 in manylinux
This also changes the way how the libssh finds the OpenSSL installation
to use the `OPENSSL_ROOT_DIR` variable.

Based partially on the changes in #719, partially on current version of
https://github.com/pyca/infra/tree/main/cryptography-linux

The s390x issue was fixed with the suggestion from openssl/openssl#27323

The new openssl build takes longer (especially on ppc64le) so bumping
the timeout to 50 minutes.

Signed-off-by: Jakub Jelen <jjelen@redhat.com>